+++ OrderApi/Core/OrderApi.Application/Features/Meditor/Handlers/OrderHandlers/RemoveOrderCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using OrderApi.Application.Exceptions;
 using OrderApi.Application.Features.Meditor.Command.OrderCommands;
 using OrderApi.Application.Interfaces;
 using OrderApi.Domain.Entities;
@@ -22,8 +23,14 @@ namespace OrderApi.Application.Features.Meditor.Handlers.OrderHandlers
         public async Task Handle(RemoveOrderCommand request, CancellationToken cancellationToken)
         {
             var values = await _repository.GetByIdAsync(request.Id);
-            _repository.Delete(values);
-            await _unitOfWork.Commit();
+
+            if(values != null)
+            {
+
+                _repository.Delete(values);
+                await _unitOfWork.Commit();
+            }
+            throw new NotFoundIdException(request.Id);
         }
     }
 }
